integer size optimization for gas fee reduction

+++ contracts/PianoKingVote.sol
@@ -11,7 +11,7 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
   using Address for address payable;
 
   // Store tokens who were already used for current vote
-  mapping(uint256 => bool) private usedTokens;
+  mapping(uint16 => bool) private usedTokens;
 
   //if a vote is open
   bool private voteOpen = false;
@@ -20,7 +20,7 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
   address[] private choices;
 
   //storage for current votes
-  uint256[] private votes;
+  uint8[] private votes;
  
   //pianoking smart contract
   PianoKing public pianoKing;
@@ -32,7 +32,7 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
   /**
    * @dev vote for a given choice with given piano king tokens sender owns
    */
-  function vote(uint256[] memory tokenIds, uint256 choice) external {
+  function vote(uint16[] memory tokenIds, uint8 choice) external {
     // Checks there is an ongoing vote
     require(voteOpen, "No ongoing vote");
     // Checks tokenIds size doesn't exceed limit
@@ -42,12 +42,12 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
     // Checks if choice is correct
     require(choice>=0&&choice<choices.length, "this choice doesn't exist");
     // Checks if tokens are not already used
-    for(uint256 i=0;i < tokenIds.length;i++){
+    for(uint8 i=0;i < tokenIds.length;i++){
       require(!usedTokens[tokenIds[i]], "one token was already used");
       require(pianoKing.ownerOf(tokenIds[i])==msg.sender, "one token is not yours");
     }
-    votes[choice] += tokenIds.length;
-    for(uint256 i=0;i < tokenIds.length;i++){
+    votes[choice] += uint16(tokenIds.length);
+    for(uint8 i=0;i < tokenIds.length;i++){
       usedTokens[tokenIds[i]] = true;
     }
   }
@@ -61,7 +61,7 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
   /**
    * @dev Gets the vote array
    */
-  function getVotes() external view returns (uint256[] memory) {
+  function getVotes() external view returns (uint16[] memory) {
     return votes;
   }
 
@@ -86,12 +86,12 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
     require(!voteOpen, "Already ongoing vote");
     require(_choices.length>0, "At least 1 choice");
     require(_choices.length<10, "max 10 choices");
-    for(uint256 i=0;i < _choices.length;i++){
+    for(uint8 i=0;i < _choices.length;i++){
       require(_choices[i] != address(0), "Invalid address");
     }
     choices = _choices;
-    votes = new uint256[](_choices.length);
-    for(uint256 i=1;i<pianoKing.totalSupply()+1;i++){
+    votes = new uint16[](_choices.length);
+    for(uint16 i=1;i<pianoKing.totalSupply()+1;i++){
           delete(usedTokens[i]);
     }
     voteOpen = true;
@@ -102,14 +102,14 @@ contract PianoKingVote is Ownable, ReentrancyGuard {
    */
   function endVote() external onlyOwner {
     require(voteOpen, "No ongoing vote");
-    uint256 total = 0;
-    for(uint256 i=0;i<choices.length;i++){
+    uint16 total = 0;
+    for(uint8 i=0;i<choices.length;i++){
       total+=votes[i];
     }
     require(total>0&&pianoKing.totalSupply()/total<=10, "Not enough votes");
-    uint256 maxVote = votes[0];
-    uint256 index = 0;
-    for(uint256 i=1;i<choices.length;i++){
+    uint16 maxVote = votes[0];
+    uint8 index = 0;
+    for(uint8 i=1;i<choices.length;i++){
       if(votes[i]>maxVote){
         maxVote = votes[i];
         index = i;
